+++ Core/Src/run.c
@@ -8,16 +8,11 @@
 
 #include "run.h"
 #include "bno085_runner.h"
-#include "can_id_allocatee.h"
 
 /** Public functions. *********************************************************/
 
 void momentum_run(void) {
   bno085_run(); // BNO085 process.
-
-  // TODO: Consider some kind of conditional flagged operation to reduce
-  //  overhead.
-  can_id_allocatee_state_machine(); // CAN ID allocatee state machine.
 
   scheduler_run(); // Run the scheduler.
 }